personnalisation & bug fix
personnalisation PID, BED,custom menu, bug fix PID bed

+++ Marlin/temperature.cpp
@@ -240,7 +240,8 @@ uint8_t Temperature::soft_pwm_amount[HOTENDS],
           (WATCH_TEMP_INCREASE + TEMP_HYSTERESIS + 1)
         #endif
       ;
-      const int8_t watch_temp_period =
+      //const int8_t watch_temp_period =
+       const uint16_t watch_temp_period =
         #if ENABLED(THERMAL_PROTECTION_BED) && ENABLED(PIDTEMPBED) && ENABLED(THERMAL_PROTECTION_HOTENDS) && ENABLED(PIDTEMP)
           hotend < 0 ? WATCH_BED_TEMP_PERIOD : WATCH_TEMP_PERIOD
         #elif ENABLED(THERMAL_PROTECTION_BED) && ENABLED(PIDTEMPBED)
@@ -249,7 +250,7 @@ uint8_t Temperature::soft_pwm_amount[HOTENDS],
           WATCH_TEMP_PERIOD
         #endif
       ;
-      const int8_t watch_temp_increase =
+      const uint8_t watch_temp_increase =
         #if ENABLED(THERMAL_PROTECTION_BED) && ENABLED(PIDTEMPBED) && ENABLED(THERMAL_PROTECTION_HOTENDS) && ENABLED(PIDTEMP)
           hotend < 0 ? WATCH_BED_TEMP_INCREASE : WATCH_TEMP_INCREASE
         #elif ENABLED(THERMAL_PROTECTION_BED) && ENABLED(PIDTEMPBED)
